+++ tests/unit/CRUD/odata/odata-reading-custom-predicates-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-odata-test';
+import readingPredicatesCustomPredicates from '../base/base-reading-custom-predicates-test';
+
+executeTest('reading | predicates | custom predicates', (store, assert) => {
+  readingPredicatesCustomPredicates(store, assert);
+});